feat: 数量字段支持三位小数 + 时区修复 + 期初库存入口
后端：
- 数量字段从 Integer 改为 BigDecimal，支持三位小数（实体类、VO、Mapper、Service）
- StockReportItemVO/StockValueReportItemVO 数量字段改 BigDecimal，修复库存报表小数被截断
- FinStockLedgerMapper.xml resultType 从 Integer 改为 BigDecimal
- ApplicationConfig 设置 JVM 默认时区 GMT+8，BaseEntity 字段添加 @JsonFormat(timezone=GMT+8)
- FinSaleRecordController 缴款时间保存完整时间戳，SimpleDateFormat 设置时区
- FinSalePayment/FinSaleRecord 日期字段添加 @JsonFormat(timezone=GMT+8)
- 进货单 dept_id 保留原值，admin 跨部门编辑不再覆盖

前端PC：
- 金额统一两位小数显示（money() 全局工具函数）
- sale/purchase 数量输入 el-input-number 添加 precision=3
- constantRoutes 期初库存入口 hidden 改为 false，侧边栏显示

小程序：
- trimNumber 改用 toFixed(3)，form/index.vue 新增 toNum3 支持小数输入

SQL：
- 新增 2026-07-29-quantity-decimal-migration.sql，21 列 INT→DECIMAL(18,3)

部署脚本：
- deploy-common.sh 编译参数改为 -Dmaven.test.skip=true

+++ junsong-modules/junsong-finance/src/main/java/com/junsong/finance/domain/FinPurchase.java
@@ -58,7 +58,7 @@ public class FinPurchase extends BaseEntity
 
     /** 总数量（含赠品） */
     @Excel(name = "总数量", cellType = ColumnType.NUMERIC)
-    private Integer totalQuantity;
+    private BigDecimal totalQuantity;
 
     /** 状态（0草稿 1已确认 2已完成） */
     @Excel(name = "状态", readConverterExp = "0=草稿,1=已确认,2=已完成")
@@ -179,12 +179,12 @@ public class FinPurchase extends BaseEntity
         this.paymentMethod = paymentMethod;
     }
 
-    public Integer getTotalQuantity()
+    public BigDecimal getTotalQuantity()
     {
         return totalQuantity;
     }
 
-    public void setTotalQuantity(Integer totalQuantity)
+    public void setTotalQuantity(BigDecimal totalQuantity)
     {
         this.totalQuantity = totalQuantity;
     }
