+++ junsong-modules/junsong-finance/src/main/java/com/junsong/finance/domain/vo/FinStockPositionView.java
@@ -1,5 +1,7 @@
 package com.junsong.finance.domain.vo;
 
+import java.math.BigDecimal;
+
 /**
  * 库存当前结存视图：用于快照生成时读取 fin_stock_position 的轻量投影。
  *
@@ -10,7 +12,7 @@ public class FinStockPositionView {
     private Long tenantId;
     private Long deptId;
     private Long productId;
-    private Integer quantity;
+    private BigDecimal quantity;
     private String productName;
 
     public Long getTenantId() { return tenantId; }
@@ -22,8 +24,8 @@ public class FinStockPositionView {
     public Long getProductId() { return productId; }
     public void setProductId(Long productId) { this.productId = productId; }
 
-    public Integer getQuantity() { return quantity; }
-    public void setQuantity(Integer quantity) { this.quantity = quantity; }
+    public BigDecimal getQuantity() { return quantity; }
+    public void setQuantity(BigDecimal quantity) { this.quantity = quantity; }
 
     public String getProductName() { return productName; }
     public void setProductName(String productName) { this.productName = productName; }
